review validation works on backend, still working on frontend validation

+++ app/api/review_routes.py
@@ -5,7 +5,6 @@ import datetime
 from flask_login import login_required, current_user
 review_routes = Blueprint('reviews', __name__)
 
-
 # get all reviews, mainly for testing
 @review_routes.route('/')
 def allReviews():
@@ -14,42 +13,27 @@ def allReviews():
     for review in reviews:
         print(review.to_dict())
         reviewList.append(review.to_dict())
-    # reviewsDict = [reviews.to_dict() for review in reviews]
     return {'reviews': reviewList}
-
-# get reviews by product id
-# @review_routes.route('/<int:id>')
-# def productReviews(id):
-#     product = Product.query.get(id)
-#     print("PRODUCT", product)
-#     pd = product.to_dict()
-#     user = product.users.to_dict()
-#     reviews = product.reviews
-#     pdReviews = {'reviews': [reviews.to_dict() for reviews in reviews]}
-#     pdUser = {'seller': user}
-#     pd.update(pdUser)
-#     pd.update(pdReviews)
-#     return pd
-
-# get reviews by user
-
 
 # create a review
 @review_routes.route('/', methods=['POST'])
 @login_required
 def createReview():
     date = datetime.datetime.now()
-    print("DATE", date)
     data = request.get_json()
     form = ReviewForm()
     form['csrf_token'].data = request.cookies['csrf_token'] # makes a csrf_token in form object
+
     if form.validate_on_submit():
+        old_review = Review.query.filter_by(product_id=data["product_id"], user_id=current_user.id).first()
+        if old_review:
+            return {"message": "You have already reviewed this product. You can't submit another review"}, 400
         new_review = Review(
             review = data["review"],
             rating = data["rating"],
             product_id = data["product_id"],
-            user_id = data["user_id"],
-            created_at = date
+            user_id=current_user.id,
+            created_at=date
         )
         print(new_review.to_dict())
         db.session.add(new_review)
@@ -57,15 +41,24 @@ def createReview():
 
         return new_review.to_dict()
     else:
-        return "Bad data, try again", 404
+        return {"message": "Bad data, try again"}, 400
+
 
 
 @review_routes.route('/<int:id>', methods=['PUT'])
+@login_required
 def editReview(id):
     date = datetime.datetime.now()
     data = request.get_json()
     form = ReviewForm()
-    review = Review.query.filter_by(id=id, user_id=current_user.id).first_or_404()
+    review = Review.query.filter_by(id=id).first()
+
+    if review is None:
+        return "Review not found", 404
+
+    if review.user_id != current_user.id:
+        return "You are no allowed to edit this review.", 401
+
     if form.validate_on_submit():
         review.review = data["review"]
         review.rating = data["rating"]
@@ -89,6 +82,9 @@ def deleteReview(id):
     review = Review.query.get(id)
     if not review:
         return ("Review not found"), 404
+
+    if review.user_id != current_user.id:
+        return {"error": "You can't delete this review."}, 401
 
     db.session.delete(review)
     db.session.commit()
